fix(db): repair failing database test
- worker_feedback_owner_or_admin_read: wrap auth.uid() in an init-plan
  subselect to satisfy the use-one-init-plan-per-statement contract
  (usage_optimization.test.sql test 11)
- auth_account_provisioning.test.sql: cast count(*) result to int for the
  pgtap is() matcher
- mock_gcash_payment.test.sql: give provisioned users display names, split
  the fixture bookings across distinct service requests to respect the
  one-active-booking constraint, run wallet assertions as the worker so
  wallet_owner RLS lets them through, and expect a no-op instead of a
  throw when cash-confirming an already-SUCCESSFUL GCash payment
- trust_pricing_and_reoffer.test.sql: assert the proof storage object is
  left for client-side deletion, matching the storage-guard contract

+++ supabase/tests/database/auth_account_provisioning.test.sql
@@ -80,7 +80,7 @@ select throws_ok(
 );
 
 select is(
-  (select count(*) from public.accounts where mobile = '+639171234567'),
+  (select count(*)::int from public.accounts where mobile = '+639171234567'),
   1,
   'the duplicate registration leaves no second account behind'
 );
